Add Other Bill

+++ src/main/java/com/ms/ware/online/solution/school/dao/billing/StuBillingMasterDaoImp.java
@@ -24,6 +24,7 @@ public class StuBillingMasterDaoImp implements StuBillingMasterDao {
     int row = 1;
     @Autowired
     private HibernateUtil util;
+
     @Override
     public List<StuBillingMaster> getAll(String hql) {
         msg = "";
@@ -192,9 +193,11 @@ public class StuBillingMasterDaoImp implements StuBillingMasterDao {
         row = 1;
         try {
             session.save(obj);
-            session.createSQLQuery("delete from voucher_detail where voucher_no='" + voucherNo + "';").executeUpdate();
-            session.createSQLQuery("update voucher set fee_receipt_no=null where voucher_no='" + voucherNo + "'").executeUpdate();
-            session.createSQLQuery("delete from voucher where voucher_no='" + voucherNo + "'").executeUpdate();
+            if (voucherNo != null) {
+                session.createSQLQuery("delete from voucher_detail where voucher_no='" + voucherNo + "';").executeUpdate();
+                session.createSQLQuery("update voucher set fee_receipt_no=null where voucher_no='" + voucherNo + "'").executeUpdate();
+                session.createSQLQuery("delete from voucher where voucher_no='" + voucherNo + "'").executeUpdate();
+            }
             session.createSQLQuery("delete from stu_billing_detail where bill_no='" + billNo + "'").executeUpdate();
             session.createSQLQuery("delete from stu_billing_master where bill_no='" + billNo + "'").executeUpdate();
             tr.commit();
